refactor(gtw): extract parse failure logging to shared utility
Move duplicate logging logic into utils/debug-log.js for reuse
across NewCommand, PrCommand, and PushCommand.
